requirements.txt
fastapi
uvicorn
pydantic
supabase
ipykernel
httpx
python-dotenv
setuptools
python-telegram-bot